# docs/swagger.yaml
openapi: 3.0.0
paths:
  /api/v1/health:
    get:
      operationId: HealthController_getHealth
      parameters: []
      responses:
        '200':
          description: Health check
          content:
            application/json:
              schema:
                type: object
                example:
                  status: alive
                properties:
                  status:
                    type: string
                    example: alive
      tags:
        - Health
  '/api/v1/translation/find/by/{uuid}':
    get:
      operationId: TranslationController_findOneByUUID
      parameters:
        - name: uuid
          required: true
          in: path
          description: Dictionary id
          schema:
            type: string
      responses:
        '200':
          description: Returns OK when successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FetchTranslationResponse'
        '400':
          description: Returns BAD_REQUEST when the payload is invalid or malformed.
          content:
            application/json:
              schema:
                type: object
                example:
                  statusCode: 400
                  message:
                    - name should not be empty
                    - name should be string
                  error: Bad Request
                properties:
                  message:
                    type: array
                  statusCode:
                    type: number
                  error:
                    type: string
        '404':
          description: Returns NOT_FOUND when no record found with the given uuid
      tags: &ref_0
        - Translation
  '/api/v1/translation/create/for/{uuid}':
    post:
      operationId: TranslationController_create
      parameters:
        - name: uuid
          required: true
          in: path
          example: 12345678-1234-1234-1234-123456789012
          description: UUID of the dictionary
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTranslationRequest'
      responses:
        '201':
          description: Returns CREATED when successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateDictionaryResponse'
        '400':
          description: Returns BAD_REQUEST when the payload is invalid or malformed.
          content:
            application/json:
              schema:
                type: object
                example:
                  statusCode: 400
                  message:
                    - name should not be empty
                    - name should be string
                  error: Bad Request
                properties:
                  message:
                    type: array
                  statusCode:
                    type: number
                  error:
                    type: string
        '401':
          description: 'Returns UNAUTHORIZED when JWT is not provided, invalid or expired.'
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized
        '403':
          description: Returns FORBIDDEN when JWT is valid but user is not authorized.
          content:
            application/json:
              schema:
                example:
                  statusCode: 403
                  message: Forbidden
        '422':
          description: Returns UNPROCESSABLE_ENTITY when UUID is wrong
      tags: *ref_0
      security:
        - bearer: []
  '/api/v1/translation/update/by/{uuid}':
    patch:
      operationId: TranslationController_update
      parameters:
        - name: uuid
          required: true
          in: path
          description: Dictionary UUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTranslationRequest'
      responses:
        '204':
          description: Returns NO_CONTENT when successful
        '400':
          description: Returns BAD_REQUEST when the payload is invalid or malformed.
          content:
            application/json:
              schema:
                type: object
                example:
                  statusCode: 400
                  message:
                    - name should not be empty
                    - name should be string
                  error: Bad Request
                properties:
                  message:
                    type: array
                  statusCode:
                    type: number
                  error:
                    type: string
        '401':
          description: 'Returns UNAUTHORIZED when JWT is not provided, invalid or expired.'
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized
        '403':
          description: Returns FORBIDDEN when JWT is valid but user is not authorized.
          content:
            application/json:
              schema:
                example:
                  statusCode: 403
                  message: Forbidden
        '422':
          description: Returns UNPROCESSABLE_ENTITY when UUID is wrong
      tags: *ref_0
      security:
        - bearer: []
  '/api/v1/translation/delete/by/{uuid}':
    delete:
      operationId: TranslationController_delete
      parameters:
        - name: uuid
          required: true
          in: path
          description: Dictionary UUID
          schema:
            type: string
      responses:
        '202':
          description: Returns ACCEPTED when successful
        '400':
          description: Returns BAD_REQUEST when the payload is invalid or malformed.
          content:
            application/json:
              schema:
                type: object
                example:
                  statusCode: 400
                  message:
                    - name should not be empty
                    - name should be string
                  error: Bad Request
                properties:
                  message:
                    type: array
                  statusCode:
                    type: number
                  error:
                    type: string
        '401':
          description: 'Returns UNAUTHORIZED when JWT is not provided, invalid or expired.'
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized
        '403':
          description: Returns FORBIDDEN when JWT is valid but user is not authorized.
          content:
            application/json:
              schema:
                example:
                  statusCode: 403
                  message: Forbidden
        '422':
          description: Returns UNPROCESSABLE_ENTITY when UUID is wrong
      tags: *ref_0
      security:
        - bearer: []
  '/api/v1/dictionary/find/by/{uuid}':
    get:
      operationId: DictionaryController_findOneByUUID
      parameters:
        - name: uuid
          required: true
          in: path
          description: Dictionary id
          schema:
            type: string
      responses:
        '200':
          description: Returns OK when successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FetchDictionaryResponse'
        '400':
          description: Returns BAD_REQUEST when the payload is invalid or malformed.
          content:
            application/json:
              schema:
                type: object
                example:
                  statusCode: 400
                  message:
                    - name should not be empty
                    - name should be string
                  error: Bad Request
                properties:
                  message:
                    type: array
                  statusCode:
                    type: number
                  error:
                    type: string
      tags: &ref_1
        - Dictionary
  /api/v1/dictionary/create:
    post:
      operationId: DictionaryController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDictionaryRequest'
      responses:
        '201':
          description: Returns CREATED when successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateDictionaryResponse'
        '400':
          description: Returns BAD_REQUEST when the payload is invalid or malformed.
          content:
            application/json:
              schema:
                type: object
                example:
                  statusCode: 400
                  message:
                    - name should not be empty
                    - name should be string
                  error: Bad Request
                properties:
                  message:
                    type: array
                  statusCode:
                    type: number
                  error:
                    type: string
        '401':
          description: 'Returns UNAUTHORIZED when JWT is not provided, invalid or expired.'
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized
        '403':
          description: Returns FORBIDDEN when JWT is valid but user is not authorized.
          content:
            application/json:
              schema:
                example:
                  statusCode: 403
                  message: Forbidden
      tags: *ref_1
      security:
        - bearer: []
  '/api/v1/dictionary/update/by/{uuid}':
    patch:
      operationId: DictionaryController_update
      parameters:
        - name: uuid
          required: true
          in: path
          description: Dictionary UUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDictionaryRequest'
      responses:
        '204':
          description: Returns NO_CONTENT when successful
        '400':
          description: Returns BAD_REQUEST when the payload is invalid or malformed.
          content:
            application/json:
              schema:
                type: object
                example:
                  statusCode: 400
                  message:
                    - name should not be empty
                    - name should be string
                  error: Bad Request
                properties:
                  message:
                    type: array
                  statusCode:
                    type: number
                  error:
                    type: string
        '401':
          description: 'Returns UNAUTHORIZED when JWT is not provided, invalid or expired.'
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized
        '403':
          description: Returns FORBIDDEN when JWT is valid but user is not authorized.
          content:
            application/json:
              schema:
                example:
                  statusCode: 403
                  message: Forbidden
        '422':
          description: Returns UNPROCESSABLE_ENTITY when UUID is wrong
      tags: *ref_1
      security:
        - bearer: []
  '/api/v1/dictionary/delete/by/{uuid}':
    delete:
      operationId: DictionaryController_delete
      parameters:
        - name: uuid
          required: true
          in: path
          description: Dictionary UUID
          schema:
            type: string
      responses:
        '202':
          description: Returns ACCEPTED when successful
        '400':
          description: Returns BAD_REQUEST when the payload is invalid or malformed.
          content:
            application/json:
              schema:
                type: object
                example:
                  statusCode: 400
                  message:
                    - name should not be empty
                    - name should be string
                  error: Bad Request
                properties:
                  message:
                    type: array
                  statusCode:
                    type: number
                  error:
                    type: string
        '401':
          description: 'Returns UNAUTHORIZED when JWT is not provided, invalid or expired.'
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized
        '403':
          description: Returns FORBIDDEN when JWT is valid but user is not authorized.
          content:
            application/json:
              schema:
                example:
                  statusCode: 403
                  message: Forbidden
        '422':
          description: Returns UNPROCESSABLE_ENTITY when UUID is wrong
      tags: *ref_1
      security:
        - bearer: []
  '/api/v1/search/search/by/{keyword}/{language}':
    get:
      operationId: SearchController_searchByKeyword
      parameters:
        - name: keyword
          required: true
          in: path
          description: What keyword to use
          schema:
            type: string
        - name: language
          required: true
          in: path
          description: What language to use
          schema:
            enum:
              - en
              - ru
              - ka
              - me
              - sv
            type: string
      responses:
        '200':
          description: Returns OK when successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FetchDictionaryResponse'
        '400':
          description: Returns BAD_REQUEST when the payload is invalid or malformed.
          content:
            application/json:
              schema:
                type: object
                example:
                  statusCode: 400
                  message:
                    - name should not be empty
                    - name should be string
                  error: Bad Request
                properties:
                  message:
                    type: array
                  statusCode:
                    type: number
                  error:
                    type: string
      tags:
        - Search
info:
  title: Megrulad.ge Backend Service
  description: Main translation backend service
  version: 0.0.3
  contact: {}
tags: []
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Enter JWT auth token
      in: header
  schemas:
    FetchTranslationResponse:
      type: object
      properties:
        uuid:
          type: string
          description: UUID string
          example: 12345678-1234-1234-1234-123456789012
        value:
          type: string
          example: Hello there
          description: Word or sentence
        description:
          type: string
          example: Some explanation
          description: 'Dictionary word/sentence description, with examples'
        language:
          type: string
          example: en
          description: Language code
      required:
        - uuid
        - value
        - description
        - language
    CreateTranslationRequest:
      type: object
      properties:
        value:
          type: string
          example: Explanation
          description: Word or sentence
        description:
          type: string
          example: Some explanation
          description: 'Translation word/sentence description, with examples'
        language:
          type: string
          example: en
          description: Language code
          enum: &ref_2
            - en
            - ru
            - ka
            - me
            - sv
        source:
          type: string
          description: >-
            Source of the word/sentence, whether it's individually contributed
            by user, or from a dictionary
          default: null
      required:
        - value
        - language
    CreateDictionaryResponse:
      type: object
      properties:
        uuid:
          type: string
          description: UUID string
          example: 12345678-1234-1234-1234-123456789012
        value:
          type: string
          example: Hello there
          description: Word or sentence
        description:
          type: string
          example: Some explanation
          description: 'Dictionary word/sentence description, with examples'
        language:
          type: string
          example: en
          description: Language code
      required:
        - uuid
        - value
        - description
        - language
    UpdateTranslationRequest:
      type: object
      properties:
        value:
          type: string
          example: Explanation
          description: Word or sentence
        description:
          type: string
          example: Some explanation
          description: 'Translation word/sentence description, with examples'
        language:
          type: string
          example: en
          description: Language code
          enum: *ref_2
        source:
          type: string
          description: >-
            Source of the word/sentence, whether it's individually contributed
            by user, or from a dictionary
          default: null
        status:
          type: string
          description: Record status
          enum:
            - ACTIVE
            - INACTIVE
            - DELETED
            - PENDING
            - REJECTED
            - HIDDEN
    TranslationProperty:
      type: object
      properties:
        uuid:
          type: string
          description: UUID string
          example: 12345678-1234-1234-1234-123456789012
        value:
          type: string
          example: Hello there
          description: Word or sentence
        description:
          type: string
          example: Some explanation
          description: 'Translation word/sentence description, with examples'
        language:
          type: string
          example: en
          description: Language code
      required:
        - uuid
        - value
        - description
        - language
    FetchDictionaryResponse:
      type: object
      properties:
        uuid:
          type: string
          description: UUID string
          example: 12345678-1234-1234-1234-123456789012
        value:
          type: string
          example: Hello there
          description: Word or sentence
        description:
          type: string
          example: Some explanation
          description: 'Dictionary word/sentence description, with examples'
        language:
          type: string
          example: en
          description: Language code
        translations:
          type: array
          items:
            $ref: '#/components/schemas/TranslationProperty'
      required:
        - uuid
        - value
        - description
        - language
        - translations
    CreateDictionaryRequest:
      type: object
      properties:
        value:
          type: string
          example: Hello there
          description: Word or sentence
        description:
          type: string
          example: Some explanation
          description: 'Dictionary word/sentence description, with examples'
        language:
          type: string
          example: en
          description: Language code
          enum: &ref_3
            - en
            - ru
            - ka
            - me
            - sv
        source:
          type: string
          description: >-
            Source of the word/sentence, whether it's individually contributed
            by user, or from a dictionary
          default: null
      required:
        - value
        - language
    UpdateDictionaryRequest:
      type: object
      properties:
        value:
          type: string
          example: Hello there
          description: Word or sentence
        description:
          type: string
          example: Some explanation
          description: 'Dictionary word/sentence description, with examples'
        language:
          type: string
          example: en
          description: Language code
          enum: *ref_3
        source:
          type: string
          description: >-
            Source of the word/sentence, whether it's individually contributed
            by user, or from a dictionary
          default: null
        status:
          type: string
          description: Record status
          enum:
            - ACTIVE
            - INACTIVE
            - DELETED
            - PENDING
            - REJECTED
            - HIDDEN